+++ app/src/main/java/com/example/seminar_assignment_2025/data/dto/GenreDto.kt
@@ -1,0 +1,9 @@
+package com.example.seminar_assignment_2025.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GenreDto(
+    val id: Int,
+    val name: String
+)
