+++ app/src/main/java/com/example/seminar_assignment_2025/data/dto/MovieSearchResponse.kt
@@ -1,6 +1,5 @@
 package com.example.seminar_assignment_2025.data.dto
 
-import com.example.seminar_assignment_2025.data.MovieDto
 import kotlinx.serialization.Serializable
 
 @Serializable
